move readonlyiterator
