+++ src/templater/language/Element.java
@@ -5,8 +5,6 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
-import templater.ReadOnlyIterator;
-
 /**
  * A class representing a {@code Element}, an HTML element.
  *
